+++ src/pydvl/value/shapley/__init__.py
@@ -8,7 +8,7 @@ Please refer to :ref:`data valuation` for an overview of Shapley Data value.
 from typing import Optional, cast
 
 from pydvl.utils import Utility
-from pydvl.value.convergence import StoppingCriterion, max_iterations
+from pydvl.value.stopping import StoppingCriterion, max_iterations
 from pydvl.value.results import ValuationResult
 from pydvl.value.shapley.gt import group_testing_shapley
 from pydvl.value.shapley.knn import knn_shapley
